using goresp v0.1.8

+++ godis/handler.go
@@ -64,7 +64,7 @@ func del(args []Value) Value {
 	}
 	value := MemDbInstance.sets.Del(keys)
 
-	return Value{Typ: "int", Num: value}
+	return Value{Typ: "int", Num: int64(value)}
 
 }
 
@@ -81,7 +81,7 @@ func expire(args []Value) Value {
 	if res == -1 {
 		return goresp.NewErrorValue(fmt.Sprintf("Invalid Option for Expire: option %s ", args[2].Bulk))
 	}
-	return goresp.NewNumberValue(int16(res))
+	return goresp.NewNumberValue(int64(res))
 }
 
 func hset(args []Value) Value {
